Montandolo en la net

+++ database/migrations/2025_09_24_110042_create_team_invitations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('sender_id');
             $table->unsignedBigInteger('receiver_id');
-            $table->unsignedInteger('team_id');
+            $table->unsignedBigInteger('team_id');
 
             $table->enum('status', ['pending', 'accepted', 'rejected'])->default('pending');
             $table->string('type')->default('invite'); // invite o response
